backend/requirements.txt
boto3==1.9.84
braintree==3.49.0
certifi==2017.11.5
chardet==3.0.4
marshmallow==2.20.2
click==7.1.2
coreapi==2.3.3
Django==2.0.5
django-choices==1.6.0
django-cors-headers==2.2.0
django-countries==5.0
django-filter==2.1.0
django-localflavor==1.6.2
django-money==2.0
django-redis==4.9.0
djangorestframework==3.8.2
django-storages==1.7.1
drf-extensions==0.3.1
djangorestframework-camel-case==0.2.0
djangorestframework-jwt==1.11.0
djangorestframework-xml==2.0.0
facebook_business==15.0.0
factory-boy==2.12.0
django-rest-swagger==2.2.0
graphene-django==2.3.2
hashids==1.2.0
html2text==2018.1.9
idna==2.6
ipython==7.16.3
klaviyo==3.0.0
mock==4.0.3
optimizely-sdk==2.1.1
Pillow==8.0.0
psycopg2-binary
pycountry>=19.8.18
pygments==2.2.0
PyJWT==1.5.3
pytz==2017.3
requests==2.20.0
six==1.12.0
stripe==2.55.0
taxjar==2.0.0
urllib3==1.22
whitenoise==3.3.1
django-ipware==2.1.0
pyotp==2.6.0
analytics-python==1.2.9
django-test-without-migrations==0.5
django_extensions==3.0.3
celery==5.1.0
redis==3.5.3
ShopifyAPI
xmltodict==0.12.0
sentry-sdk==1.6.0
